+++ src/index.html
@@ -5,7 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://kit.fontawesome.com/51f27b15da.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="style.css">
     <title>Document</title>
 </head>
 <body>
@@ -23,6 +22,5 @@
         <ul id="list"></ul>
         <a href="#">Clear all completed</a>
     </div>
-    <script type="module" src="index.js"></script>
 </body>
 </html>
